+++ src/MockHttp/MediaTypes.cs
@@ -5,4 +5,5 @@ internal static class MediaTypes
     internal const string TextPlain = "text/plain";
     internal const string OctetStream = "application/octet-stream";
     internal const string DefaultMediaType = TextPlain;
+    internal const string Xml = "application/xml";
 }